Revert "Workaround attempt 1 for httpauth type writing file as root."
This reverts commit e958c4d0c4ba8b66692fe23754d22699fc8d2114.

This workaround does not work because there ends up being two file resources
with the same path. Another workaround would be to use an exec with some kind of
unless test. That's pretty awful, though, so it's probably better to go forth
and update the type to support owner/group/mode. :-\